requirements.txt
amqp==5.2.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==2.4.1
billiard==4.2.0
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==42.0.8
cssselect2==0.7.0
Django==5.0.7
django-allauth>=0.63,<1
executing==2.0.1
html5lib==1.1
icecream==2.1.3
idna==3.7
kombu==5.3.7
lxml==5.2.2
oscrypto==1.3.0
pdf2image==1.17.0
pillow==10.4.0
prompt_toolkit==3.0.47
pycairo==1.26.1
pycparser==2.22
Pygments==2.18.0
pyHanko==0.25.0
pyhanko-certvalidator==0.26.3
pypdf==4.2.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.7
reportlab==4.2.2
requests==2.32.3
rq==1.16.2
six==1.16.0
sqlparse==0.5.0
svglib==1.5.1
tinycss2==1.3.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
xhtml2pdf==0.2.16
